+++ main.tf
@@ -521,7 +521,7 @@ data "aws_iam_policy_document" "combined" {
     var.attach_lb_log_delivery_policy ? data.aws_iam_policy_document.lb_log_delivery[0].json : "",
     var.attach_require_latest_tls_policy ? data.aws_iam_policy_document.require_latest_tls[0].json : "",
     var.attach_deny_insecure_transport_policy ? data.aws_iam_policy_document.deny_insecure_transport[0].json : "",
-    var.attach_inventory_destination_policy ? data.aws_iam_policy_document.inventory_destination_policy[0].json : "",
+    var.attach_inventory_destination_policy || var.attach_analytics_destination_policy ? data.aws_iam_policy_document.inventory_and_analytics_destination_policy[0].json : "",
     var.attach_policy ? var.policy : ""
   ])
 }
@@ -534,12 +534,33 @@ data "aws_elb_service_account" "this" {
 data "aws_iam_policy_document" "elb_log_delivery" {
   count = local.create_bucket && var.attach_elb_log_delivery_policy ? 1 : 0
 
+  # Policy for AWS Regions created before August 2022 (e.g. US East (N. Virginia), Asia Pacific (Singapore), Asia Pacific (Sydney), Asia Pacific (Tokyo), Europe (Ireland))
   statement {
     sid = ""
 
     principals {
       type        = "AWS"
       identifiers = data.aws_elb_service_account.this[*].arn
+    }
+
+    effect = "Allow"
+
+    actions = [
+      "s3:PutObject",
+    ]
+
+    resources = [
+      "${aws_s3_bucket.this[0].arn}/*",
+    ]
+  }
+
+  # Policy for AWS Regions created after August 2022 (e.g. Asia Pacific (Hyderabad), Asia Pacific (Melbourne), Europe (Spain), Europe (Zurich), Middle East (UAE))
+  statement {
+    sid = ""
+
+    principals {
+      type        = "Service"
+      identifiers = ["logdelivery.elasticloadbalancing.amazonaws.com"]
     }
 
     effect = "Allow"
@@ -795,13 +816,13 @@ resource "aws_s3_bucket_inventory" "this" {
   }
 }
 
-# Inventory destination bucket requires a bucket policy to allow source to PutObjects
+# Inventory and analytics destination bucket requires a bucket policy to allow source to PutObjects
 # https://docs.aws.amazon.com/AmazonS3/latest/userguide/example-bucket-policies.html#example-bucket-policies-use-case-9
-data "aws_iam_policy_document" "inventory_destination_policy" {
-  count = local.create_bucket && var.attach_inventory_destination_policy ? 1 : 0
+data "aws_iam_policy_document" "inventory_and_analytics_destination_policy" {
+  count = local.create_bucket && var.attach_inventory_destination_policy || var.attach_analytics_destination_policy ? 1 : 0
 
   statement {
-    sid    = "destinationInventoryPolicy"
+    sid    = "destinationInventoryAndAnalyticsPolicy"
     effect = "Allow"
 
     actions = [
@@ -820,16 +841,18 @@ data "aws_iam_policy_document" "inventory_destination_policy" {
     condition {
       test     = "ArnLike"
       variable = "aws:SourceArn"
-      values = [
-        var.inventory_self_source_destination ? aws_s3_bucket.this[0].arn : var.inventory_source_bucket_arn
-      ]
+      values = compact(distinct([
+        var.inventory_self_source_destination ? aws_s3_bucket.this[0].arn : var.inventory_source_bucket_arn,
+        var.analytics_self_source_destination ? aws_s3_bucket.this[0].arn : var.analytics_source_bucket_arn
+      ]))
     }
 
     condition {
       test = "StringEquals"
-      values = [
-        var.inventory_self_source_destination ? data.aws_caller_identity.current.id : var.inventory_source_account_id
-      ]
+      values = compact(distinct([
+        var.inventory_self_source_destination ? data.aws_caller_identity.current.id : var.inventory_source_account_id,
+        var.analytics_self_source_destination ? data.aws_caller_identity.current.id : var.analytics_source_account_id
+      ]))
       variable = "aws:SourceAccount"
     }
 
@@ -837,6 +860,43 @@ data "aws_iam_policy_document" "inventory_destination_policy" {
       test     = "StringEquals"
       values   = ["bucket-owner-full-control"]
       variable = "s3:x-amz-acl"
+    }
+  }
+}
+
+resource "aws_s3_bucket_analytics_configuration" "this" {
+  for_each = { for k, v in var.analytics_configuration : k => v if local.create_bucket }
+
+  bucket = aws_s3_bucket.this[0].id
+  name   = each.key
+
+  dynamic "filter" {
+    for_each = length(try(flatten([each.value.filter]), [])) == 0 ? [] : [true]
+
+    content {
+      prefix = try(each.value.filter.prefix, null)
+      tags   = try(each.value.filter.tags, null)
+    }
+  }
+
+  dynamic "storage_class_analysis" {
+    for_each = length(try(flatten([each.value.storage_class_analysis]), [])) == 0 ? [] : [true]
+
+    content {
+
+      data_export {
+        output_schema_version = try(each.value.storage_class_analysis.output_schema_version, null)
+
+        destination {
+
+          s3_bucket_destination {
+            bucket_arn        = try(each.value.storage_class_analysis.destination_bucket_arn, aws_s3_bucket.this[0].arn)
+            bucket_account_id = try(each.value.storage_class_analysis.destination_account_id, data.aws_caller_identity.current.id)
+            format            = try(each.value.storage_class_analysis.export_format, "CSV")
+            prefix            = try(each.value.storage_class_analysis.export_prefix, null)
+          }
+        }
+      }
     }
   }
 }